also rename test module
