+++ crates/astria-sequencer-relayer/tests/it/main.rs
@@ -1,4 +1,4 @@
-mod da;
+mod data_availability;
 mod sequencer;
 mod sequencer_block;
 mod transaction;
